+++ compiler-core/src/javascript/tests/case.rs
@@ -943,3 +943,28 @@ pub fn go() {
 }"#
     )
 }
+
+// https://github.com/gleam-lang/gleam/issues/5400
+#[test]
+fn case_matching_single_character_prefixes() {
+    assert_js!(
+        r#"
+pub fn parse_digit(s: String) -> Int {
+  case s {
+    "0" <> s -> 0
+    "1" <> s -> 1
+    "2" <> s -> 2
+    "3" <> s -> 3
+    "4" <> s -> 4
+    "5" <> s -> 5
+    "6" <> s -> 6
+    "7" <> s -> 7
+    "8" <> s -> 8
+    "9" <> s -> 9
+    "" -> todo
+    _ -> panic
+  }
+}
+"#
+    );
+}